Structs

Index

MolHandler.Atom

MolHandler.Attribute

MolHandler.Coordinate

MolHandler.Frame

MolHandler.Trajectory

MolHandler.Trajectory — Type

This struct corresponding to MD trajectory. This contains 4 fields like below.

coordinates : Matrix{Coordinate}

attributes : Vector{Attribute}

natom : Number of atoms.

nframe : Number of frames in the trajectory.

The column of coordinates matrix means one snapshot. The row of coordinates matrix means time series of one atom.

[ a d g j
  b e h k
  c f i l ]

In this case, one snapshot correspond to [a b c].

MolHandler.Coordinate — Type

This correspond to coordinate of atom. This contain 3 fields like below.

x <: Real

y <: Real

z <: Real

Some operators and functions were overloaded to make Coordinate objects broadcastable.

MolHandler.Attribute — Type

This struct mean all atomic information other than coordinates. This contains 5 fields like below.

resname : Union{String, Nothing}

resid : Union{Int64, Nothing}

atomname : Union{String, Nothing}

atomid : Union{Int64, Nothing}

mass : Union{Float32, Nothing}

MolHandler.Frame — Type

This struct correspond to one snapshot of trajectory. This contains 3 fields like below.

coordinates : Vector{Coordinate}

attribute : Vector{Attribute}

natom : Number of atoms.

MolHandler.Atom — Type

This struct have all information of corresponding atom. This contains 2 fields like below.

coordinate : Reference to Coordinate object.

attribute : Reference to Attribute object.