Functions

Index

MolHandler.atom_mass

MolHandler.center_of_mass

MolHandler.clip_trajectory

MolHandler.contact_bool_matrix

MolHandler.contact_bool_matrix_parallel

MolHandler.contact_probability_matrix

MolHandler.contact_probability_matrix_parallel

MolHandler.get_atom

MolHandler.get_frame

MolHandler.pair_length_matrix

MolHandler.pair_length_matrix_parallel

MolHandler.radius_of_gyration

MolHandler.read_dcd

MolHandler.read_pdb

MolHandler.write_dcd

MolHandler.write_pdb

File IO

MolHandler.read_dcd — Function

read_dcd(filename::String)::Trajectory

Return Trajectory object which filled coordinates, nframe, natom fields.

MolHandler.write_dcd — Function

write_dcd(filename::String, trj::Trajectory;
          save_step::Integer = 1, total_step::Integer = trj.nframe,
          unit_num::Integer  = 1,  time_step::Real = 1.0f0)

Write coordinates, nframe and natom infomation of trj to dcd file, named filename. dcd file contain other information - save interval step, unit(chain) number in the system, total step and time step of original trajectory -, and if you do not specify these parameter, these are filled with default value.

MolHandler.read_pdb — Function

read_pdb(filename::AbstractString; model = :unspecified)::Trajectory

Return Trajectory object which filled all field. If you set :AA to model field, this function set the mass of particle based on atomname field. If you set :CA to model field, this function set the mass of particcle mase on resname field.

MolHandler.write_pdb — Function

write_pdb(filename::AbstractString, trj::Trajectory;
          tempfactor::Union{RealT, Nothing} = nothing) where RealT <: Real

Write coordinates, resname, resid, atomname, atomid to filename based on trj. If you set Real value to tempfactor, all tempfactor will be field with that value.

Trajectory handling

MolHandler.get_frame — Function

get_frame(frame_idx::Int64, trajectory::Trajectory)
::Frame

Return Frame object which correspond to framd_idx frame from trajectory.

MolHandler.get_atom — Function

get_atom(query::Int64, trajectory::Trajectory)
::Vector{Atom}

Return Vector of queryth Atom object from trajectory.

get_atom(query::Int64, frame::Frame)
::Atom

Return queryth Atom object from frame.

MolHandler.clip_trajectory — Function

clip_trajectory(query::Union{Integer, Array{T, 1}, OrdinalRange}, trj::Trajectory;
                query_key::Symbol = :frame)
::Trajecotory where T <: Integer

Clip a part of trajectory you specified. Expected query key is one of the following

:frame : in this case, query specify the frame index.

:atom : in this case, query specify the atom index.

MolHandler.center_of_mass — Function

center_of_mass(query::Trajectory;
               frame_indices::Union{Array, OrdinalRange, Colon} = :,
               atom_indices::Union{Array, OrdinalRange, Colon} = :,
               geometric::Bool = false)
::Vector{Coordinate{Real}}

Calculate the center of mass of trajectory for specified frame and atom indices. If you set geometric is true, this function calculate geometric center of mass.

MolHandler.radius_of_gyration — Function

radius_of_gyration(trj::Trajectory;
                   frame_indices::Union{Array, OrdinalRange, Colon} = :,
                   atom_indices ::Union{Array, OrdinalRange, Colon} = :)
::Vector{Real}

Calculate radius of gyration over the trajectory.

MolHandler.pair_length_matrix — Function

pair_length_matrix(corrds1::Vector{Coordinate}, coords2::Vector{Coordinate})
::Matrix{Coordinate}

Calculate distance matrix for all combination between coords1 and coord2. The row of returned matrix coresspond to coords1, and the column correspond to coords2.

pair_length_matrix(trj::Trajectory;
                   frame_indices::Union{Array, OrdinalRange, Colon}       = :,
                   first_atom_indices::Union{Array, OrdinalRange, Colon}  = :,
                   second_atom_indices::Union{Array, OrdinalRange, Colon} = atom_indices1)
::Vector{Matrix{Coordinate}}

Calculate distance matrix for all combination between first_atom_indices and second_atom_indices. This cauculation apply to each frame of trajectory and the result matrices are stored in Vector. The target frame can be restricted by pass indeces vector or range to frame_indices.

MolHandler.contact_bool_matrix — Function

contact_bool_matrix(threshold::Real, trj::Trajectory;
                    frame_indices::Union{Array, OrdinalRange, Colon}       = :,
                    first_atom_indices::Union{Array, OrdinalRange, Colon}  = :,
                    second_atom_indices::Union{Array, OrdinalRange, Colon} = first_atom_indices)
::Vector{Matrix{Bool}}

Judge contact is formed or not. If the distance between two coordinate is shorter than threshold, contact is considered to be formed. In returned matrix, the row of matrices corresponds to firstatomindices and column of matrices corresponds to secondatomindices.

contact_bool_matrix(threshold::Real, trj::Trajectory;
                    frame_indices::Union{Array, OrdinalRange, Colon}       = :,
                    first_atom_indices::Union{Array, OrdinalRange, Colon}  = :,
                    second_atom_indices::Union{Array, OrdinalRange, Colon} = first_atom_indices)
::Vector{Matrix{Bool}}

Judge contact is formed or not. If the distance between two coordinate is shorter than threshold, contact is considered to be formed. In returned vector of matrices, each matrix correspond to contact matrix of each frame. You can specify the target frames or atoms by frame_indices, first_atom_indices or second_atom_indices. When you specify the target atoms, the row of matrices corresponds to firstatomindices and column of matrices corresponds to secondatomindices.

MolHandler.contact_probability_matrix — Function

contact_probability_matrix(threshold::Real, trj::Trajectory;
                           frame_indices::Union{Array, OrdinalRange, Colon}       = :,
                           first_atom_indices::Union{Array, OrdinalRange, Colon}  = :,
                           second_atom_indices::Union{Array, OrdinalRange, Colon} = first_atom_indices)
::Matrix{Real}

Calculate contact formation probability over the trajectory. If the distance between two coordinate is shorter than threshold, contact is considered to be formed. You can specify the target frames or atoms by frame_indices, first_atom_indices or second_atom_indices. When you specify the target atoms, the row of matrices corresponds to firstatomindices and column of matrices corresponds to secondatomindices.

Utility

MolHandler.atom_mass — Function

atom_mass(name::AbstractString)
::Float32

Return mass of the name atom. The kind of each atom is judged by PDB format atom names rule. https://cdn.rcsb.org/wwpdb/docs/documentation/file-format/PDBformat1992.pdf

Missing docstring.

Missing docstring for residue_mass. Check Documenter's build log for details.

Multi-Threading

MolHandler.pair_length_matrix_parallel — Function

pair_length_matrix_parallel(trj::Trajectory;
                            frame_indices::Union{Array, OrdinalRange, Colon}       = :,
                            first_atom_indices::Union{Array, OrdinalRange, Colon}  = :,
                            second_atom_indices::Union{Array, OrdinalRange, Colon} = atom_indices1)
::Vector{Matrix{Coordinate}}

Multi-threads version of pairlengthmatrix. If you set available threads number to Threads.nthreads(), this function would faster than non-parallel version.

MolHandler.contact_bool_matrix_parallel — Function

contact_bool_matrix_parallel(threshold::Real, trj::Trajectory;
                             frame_indices::Union{Array, OrdinalRange, Colon}       = :,
                             first_atom_indices::Union{Array, OrdinalRange, Colon}  = :,
                             second_atom_indices::Union{Array, OrdinalRange, Colon} = first_atom_indices)
::Vector{Matrix{Bool}}

Multi-threads version of contactboolmatrix. If you set available threads number to Threads.nthreads(), this function would faster than non-parallel version.

MolHandler.contact_probability_matrix_parallel — Function

contact_probability_matrix_parallel(threshold::Real, trj::Trajectory;
                                    frame_indices::Union{Array, OrdinalRange, Colon}       = :,
                                    first_atom_indices::Union{Array, OrdinalRange, Colon}  = :,
                                    second_atom_indices::Union{Array, OrdinalRange, Colon} = first_atom_indices)
::Matrix{Real}

Multi-threads version of contactprobabilitymatrix. If you set available threads number to Threads.nthreads(), this function would faster than non-parallel version.